+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Remove non-standard grey color from `dev` format
+
 1.1.1 / 2014-05-20
 ==================