+++ History.md
@@ -5,6 +5,7 @@ unreleased
   * Add `combined` log format
   * Add `common` log format
   * Add `morgan(format, options)` function signature
+  * Deprecate `default` format -- use `combined` format instead
   * Remove non-standard grey color from `dev` format
 
 1.1.1 / 2014-05-20