+++ History.md
@@ -1,5 +1,5 @@
-unreleased
-==========
+1.2.0 / 2014-07-19
+==================
 
   * Add `:remote-user` token
   * Add `combined` log format